fixing integer types

+++ main.cpp
@@ -7,10 +7,10 @@ int main(int argc, char **argv)
 {
     std::ifstream inf("day8.in");
     std::string l;
-    std::vector<std::vector<int>> m;
+    std::vector<std::vector<uint8_t>> m;
     while (std::getline(inf, l))
     {
-        std::vector<int> t;
+        std::vector<uint8_t> t;
         for (int i = 0; i < l.size(); i++)
         {
             t.push_back(l[i] - '0');
@@ -19,9 +19,9 @@ int main(int argc, char **argv)
     }
 
     int vt = 0;
-    for (int i = 0; i < m.size(); i++)
+    for (size_t i = 0; i < m.size(); i++)
     {
-        for (int j = 0; j < m[0].size(); j++)
+        for (size_t j = 0; j < m[0].size(); j++)
         {
             if (i == 0 || j == 0 || i == m.size() - 1 || j == m[0].size() - 1)
             {
@@ -29,7 +29,7 @@ int main(int argc, char **argv)
                 continue;
             }
             bool v = true;
-            for (int k = 0; k < i; k++)
+            for (size_t k = 0; k < i; k++)
             {
                 if (m[k][j] >= m[i][j])
                 {
@@ -43,7 +43,7 @@ int main(int argc, char **argv)
                 continue;
             }
             v = true;
-            for (int k = m.size() - 1; k > i; k--)
+            for (size_t k = m.size() - 1; k > i; k--)
             {
                 if (m[k][j] >= m[i][j])
                 {
@@ -57,7 +57,7 @@ int main(int argc, char **argv)
                 continue;
             }
             v = true;
-            for (int k = 0; k < j; k++)
+            for (size_t k = 0; k < j; k++)
             {
                 if (m[i][k] >= m[i][j])
                 {
@@ -71,7 +71,7 @@ int main(int argc, char **argv)
                 continue;
             }
             v = true;
-            for (int k = m[0].size() - 1; k > j; k--)
+            for (size_t k = m[0].size() - 1; k > j; k--)
             {
                 if (m[i][k] >= m[i][j])
                 {
@@ -85,44 +85,44 @@ int main(int argc, char **argv)
     }
     std::cout << vt << std::endl;
 
-    int mss = 0;
-    for (int i = 0; i < m.size(); i++)
+    uint32_t mss = 0;
+    for (size_t i = 0; i < m.size(); i++)
     {
-        for (int j = 0; j < m[0].size(); j++)
+        for (size_t j = 0; j < m[0].size(); j++)
         {
             if (i == 0 || j == 0 || i == m.size() - 1 || j == m[0].size() - 1)
             {
                 continue;
             }
-            int lss1 = 0;
-            for (int k = i - 1; k >= 0; k--)
+            uint32_t lss1 = 0;
+            for (int32_t k = i - 1; k >= 0; k--)
             {
                 lss1++;
                 if (m[k][j] >= m[i][j])
                     break;
             }
-            int lss2 = 0;
-            for (int k = i + 1; k < m.size(); k++)
+            uint32_t lss2 = 0;
+            for (size_t k = i + 1; k < m.size(); k++)
             {
                 lss2++;
                 if (m[k][j] >= m[i][j])
                     break;
             }
-            int lss3 = 0;
-            for (int k = j - 1; k >= 0; k--)
+            uint32_t lss3 = 0;
+            for (int32_t k = j - 1; k >= 0; k--)
             {
                 lss3++;
                 if (m[i][k] >= m[i][j])
                     break;
             }
-            int lss4 = 0;
-            for (int k = j + 1; k < m[0].size(); k++)
+            uint32_t lss4 = 0;
+            for (size_t k = j + 1; k < m[0].size(); k++)
             {
                 lss4++;
                 if (m[i][k] >= m[i][j])
                     break;
             }
-            int ss = lss1 * lss2 * lss3 * lss4;
+            uint32_t ss = lss1 * lss2 * lss3 * lss4;
             if(ss > mss)
                 mss = ss;
         }
